refactor : 불필요한 Symbol 삭제
