+++ src/main/java/chess/domain/piece/Knight.java
@@ -12,8 +12,7 @@ public class Knight extends Piece {
                 owner,
                 Score.KNIGHT,
                 Direction.knightDirections(),
-                Distance.ONE,
-                Symbol.KNIGHT
+                Distance.ONE
         );
     }
