+++ src/main/java/chess/domain/piece/Queen.java
@@ -12,8 +12,7 @@ public class Queen extends Piece {
                 owner,
                 Score.QUEEN,
                 Direction.allDirections(),
-                Distance.SEVEN,
-                Symbol.QUEEN
+                Distance.SEVEN
         );
     }
