+++ src/main/java/chess/domain/piece/Rook.java
@@ -12,8 +12,7 @@ public class Rook extends Piece {
                 owner,
                 Score.ROOK,
                 Direction.straightDirections(),
-                Distance.SEVEN,
-                Symbol.ROOK
+                Distance.SEVEN
         );
     }
